V2.7.0

+++ lib/widgets/quick_items/quick_items_widget.dart
@@ -3,13 +3,15 @@ import 'dart:async';
 
 // Flutter imports:
 import 'package:animations/animations.dart';
-import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 
 // Package imports:
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
+import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:provider/provider.dart';
+import 'package:torn_pda/models/quick_item_model.dart';
 import 'package:torn_pda/pages/quick_items/quick_items_options.dart';
+import 'package:torn_pda/providers/quick_items_faction_provider.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
 // Project imports:
@@ -17,12 +19,12 @@ import 'package:torn_pda/providers/quick_items_provider.dart';
 import 'package:torn_pda/utils/js_snippets.dart';
 
 class QuickItemsWidget extends StatefulWidget {
-  final String webviewType;
   final InAppWebViewController inAppWebViewController;
   final WebViewController webViewController;
+  final bool faction;
 
   QuickItemsWidget({
-    @required this.webviewType,
+    @required this.faction,
     this.inAppWebViewController,
     this.webViewController,
   });
@@ -33,6 +35,7 @@ class QuickItemsWidget extends StatefulWidget {
 
 class _QuickItemsWidgetState extends State<QuickItemsWidget> {
   QuickItemsProvider _itemsProvider;
+  QuickItemsProviderFaction _itemsProviderFaction;
 
   Timer _inventoryRefreshTimer;
 
@@ -51,6 +54,7 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
   @override
   Widget build(BuildContext context) {
     _itemsProvider = context.watch<QuickItemsProvider>();
+    _itemsProviderFaction = context.watch<QuickItemsProviderFaction>();
     return Padding(
       padding: const EdgeInsets.symmetric(horizontal: 5),
       child: Row(
@@ -82,8 +86,14 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
   List<Widget> _itemButtons() {
     var myList = <Widget>[];
 
-    for (var item in _itemsProvider.activeQuickItems) {
+    List<QuickItem> itemList = <QuickItem>[];
+    if (widget.faction) {
+      itemList = List.from(_itemsProviderFaction.activeQuickItemsFaction);
+    } else {
+      itemList = List.from(_itemsProvider.activeQuickItems);
+    }
 
+    for (var item in itemList) {
       Color qtyColor;
       if (item.inventory == 0) {
         qtyColor = Colors.orange[300];
@@ -93,7 +103,7 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
 
       double qtyFontSize = 12;
       var itemQty = item.inventory.toString();
-      if (!item.isLoadout) {
+      if (!item.isLoadout && !widget.faction) {
         if (item.inventory > 999 && item.inventory < 100000) {
           itemQty = "${(item.inventory / 1000).truncate().toStringAsFixed(0)}K";
         } else if (item.inventory >= 100000) {
@@ -113,46 +123,60 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
           decoration: BoxDecoration(color: Colors.grey[700]),
           child: ActionChip(
             elevation: 3,
-            side: item.isLoadout ? BorderSide(color: Colors.blue) : null,
+            side: item.isLoadout || item.isPoints ? BorderSide(color: Colors.blue) : null,
             avatar: item.isLoadout
                 ? null
-                : CircleAvatar(
-                    child: Text(
-                      itemQty,
-                      style: TextStyle(
-                        fontSize: qtyFontSize,
-                        color: qtyColor,
+                : widget.faction
+                    ? item.isPoints
+                        ? Icon(
+                            MdiIcons.alphaPCircleOutline,
+                            color: Colors.blueAccent,
+                          )
+                        : CircleAvatar(
+                            child: Image.asset(
+                              'images/icons/faction.png',
+                              width: 12,
+                              color: Colors.white,
+                            ),
+                          )
+                    : CircleAvatar(
+                        child: Text(
+                          itemQty,
+                          style: TextStyle(
+                            fontSize: qtyFontSize,
+                            color: qtyColor,
+                          ),
+                        ),
                       ),
-                    ),
-                  ),
             label: item.isLoadout
-                ? Text(item.loadoutName)
-                : item.name.split(' ').length > 1
-                    ? _splitName(item.name)
-                    : Text(
-                        item.name,
-                        softWrap: true,
-                        overflow: TextOverflow.clip,
-                        maxLines: 2,
+                ? Text(
+                    item.loadoutName,
+                    style: TextStyle(fontSize: 11),
+                  )
+                : item.isPoints
+                    ? Text(
+                        "Refill",
                         style: TextStyle(fontSize: 11),
-                      ),
+                      )
+                    : item.name.split(' ').length > 1
+                        ? _splitName(item.name.replaceAll("Blood Bag : ", "Blood: "))
+                        : Text(
+                            item.name.replaceAll("Blood Bag : ", "Blood: "),
+                            softWrap: true,
+                            overflow: TextOverflow.clip,
+                            maxLines: 2,
+                            style: TextStyle(fontSize: 11),
+                          ),
             onPressed: () async {
               if (item.isLoadout) {
-                if (widget.webviewType == "attacks") {
-                  var js = changeLoadOutJS(item: item.name.split(" ")[1], attackWebview: true);
-                  widget.webViewController.runJavascript(js);
-                } else {
-                  var js = changeLoadOutJS(item: item.name.split(" ")[1], attackWebview: false);
-                  await widget.inAppWebViewController.evaluateJavascript(source: js);
-                }
+                var js = changeLoadOutJS(item: item.name.split(" ")[1], attackWebview: false);
+                await widget.inAppWebViewController.evaluateJavascript(source: js);
               } else {
-                var js = quickItemsJS(item: item.number.toString());
-                if (widget.webviewType == "attacks") {
-                  widget.webViewController.runJavascript(js);
-                } else {
-                  await widget.inAppWebViewController.evaluateJavascript(source: js);
+                var js = quickItemsJS(item: item.number.toString(), faction: widget.faction, refill: item.isPoints);
+                await widget.inAppWebViewController.evaluateJavascript(source: js);
+                if (!widget.faction) {
+                  _itemsProvider.decreaseInventory(item);
                 }
-                _itemsProvider.decreaseInventory(item);
               }
             },
           ),
@@ -168,7 +192,9 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               Text(
-                "Configure your preferred quick items",
+                widget.faction
+                    ? "Configure your faction's armoury quick items"
+                    : "Configure your preferred quick items",
                 style: TextStyle(
                   color: Colors.orangeAccent,
                   fontSize: 12,
@@ -211,7 +237,9 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
   }
 
   _refreshInventory() {
-    _itemsProvider.updateInventoryQuantities(fullUpdate: false);
+    if (!widget.faction) {
+      _itemsProvider.updateInventoryQuantities(fullUpdate: false);
+    }
   }
 
   Widget _settingsIcon() {
@@ -221,7 +249,9 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
         transitionDuration: Duration(milliseconds: 500),
         transitionType: ContainerTransitionType.fadeThrough,
         openBuilder: (BuildContext context, VoidCallback _) {
-          return QuickItemsOptions();
+          return QuickItemsOptions(
+            isFaction: widget.faction,
+          );
         },
         closedElevation: 0,
         closedShape: const RoundedRectangleBorder(
